version 1.0.2 release
Fixed bug when plugin tries to get schedule state, gets not 200 response
code, throws exception, set the schedule as Invalid and tries to get its
state again.
Default time delay changed from 1 to 3 seconds, because 1 second
sometimes is not enough for Leaptest to create schedule state and it
returns not 200 response  code and a correct schedule is set as Invalid.